refactor(tests): extract fixed_point_u12x4_u16x16.cpp from fixed_point.cpp (part of #3130)
Fifth extraction in #3130. Moves the u12x4 + u16x16 unsigned type tests
into their own file. Both share a section because the section's u16x16
sub-section defines its own check_near() overloads and tol()/smooth_tol()
/rt_tol() helpers used only within these two type sections.

Content moved:
  - ~15 FL_TEST_CASE u12x4 tests (construction, addition, sub, mul, div,
    shifts, comparisons, math functions, advanced math, edge cases,
    precision)
  - ~30 FL_TEST_CASE u16x16 tests (default/float/from_raw construction,
    arithmetic, comparisons, mixed arithmetic, to_float, sqrt/rsqrt/pow,
    sqrt identity, floor/ceil/fract, min/max/mod, lerp/clamp/step/
    smoothstep, edge values, unsigned range, fractional precision)
  - tol(), smooth_tol(), rt_tol() constexpr helpers (only used in this
    section)
  - check_near(u16x16, ...) overloads (only used in this section)

Tests pass: bash test fl_math_fixed_point --cpp (1/1, 0.39s test time,
44.6s build).

fixed_point.cpp: 3,404 -> 2,492 LOC
fixed_point_u12x4_u16x16.cpp (new): 935 LOC

Stacked on feat/3130-fixed-point-u4x12-u8x8-extract (PR #3139).

Part of #3130. Meta tracking issue: #3127.
